+++ R/nm-validate.R
@@ -138,12 +138,12 @@ nm_validate <- function(.data, .spec, .error_on_fail = TRUE, .test_omit = NULL){
     )
 
   # MDV not equal to 1 when DV is NA
-  tests_results[["MDV not set to 1 when DV is NA"]] <-
+  tests_results[["MDV not set to 1 when DV is NA or 0"]] <-
     pass_fail(
       .code = c(
         "{arg_names$.data}",
         "dplyr::select({collapse_covs(c(flags$dv, flags$mdv))})",
-        "dplyr::filter(is.na({collapse_covs(flags$dv)}) & {collapse_covs(flags$mdv)} != 1)"
+        "dplyr::filter((is.na(as.numeric({collapse_covs(flags$dv)})) | as.numeric({collapse_covs(flags$dv)}) == 0) & {collapse_covs(flags$mdv)} != 1)"
       ),
       .required_flags = c("dv", "mdv")
     )