+++ tests/testthat/test-nm-write.R
@@ -2,8 +2,10 @@ nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
 nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".", show_col_types = FALSE)
 
 .temp_csv <- tempfile(fileext = ".csv")
+.temp_script <- tempfile(fileext = ".R")
+writeLines(text = "2 + 2", con = .temp_script)
 
-nm_write(nm, nm_spec, .temp_csv)
+nm_write(.data = nm, .spec = nm_spec, .file = .temp_csv, .this_script = .temp_script)
 .csv_in <- readr::read_csv(.temp_csv, na = ".")
 .xpt_in_name <- paste0(gsub(".csv", "", .temp_csv, fixed = TRUE),
                        "/",
@@ -26,6 +28,10 @@ test_that("nm_write write xpt: xpt data includes correct labels [NMV-NMW-002]", 
 
 test_that("nm_write works with special characters in file name [NMV-NMW-003]", {
   .temp_csv <- tempfile(fileext = "-pk.csv")
-  nm_write(nm, nm_spec, .temp_csv)
+  nm_write(.data = nm, .spec = nm_spec, .file = .temp_csv, .this_script = .temp_script)
   expect_equal(nm, readr::read_csv(.temp_csv, na = "."))
+})
+
+test_that("nm_write writes out script to source script directory [NMV-NMW-004]", {
+  expect_true(file.exists(.temp_script))
 })